server/requirements.txt
websockets>=3.2
pyyaml>=3.12
